refactor: move ICourse to structures
