+++ src/common/interfaces/structures/ICourse.ts
